[SM][DLS-5174] Moved dateFormatterForMultiDisposals logic to completionDateForm.
